Move netlinkutils library to libcalico-go
